Add a new package for unit tests
