+++ src/test/java/com/events/events/integration/user/BaseIntegrationTest.java
@@ -1,0 +1,4 @@
+package com.events.events.integration.user;
+
+public class BaseIntegrationTest {
+}
